Server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
faster-whisper>=0.10.0
python-Levenshtein>=0.23.0
numpy>=1.24.0
